display progress as double

+++ BDInfo.Core/AnotherBDInfo/ListElement.cs
@@ -6,7 +6,7 @@ namespace AnotherBDInfo
   {
     private readonly int _position;
     private string _text;
-    private int _value;
+    private double _value;
 
     public ListElement(int rowPosition)
     {
@@ -17,9 +17,9 @@ namespace AnotherBDInfo
 
     public string Text { get { return _text; } set { _text = value; OnTextChanged?.Invoke(_text, _position); } }
 
-    public int Value { get { return _value; } set { _value = value; OnProgressChanged?.Invoke(_value, _position); } }
+    public double Value { get { return _value; } set { _value = value; OnProgressChanged?.Invoke(_value, _position); } }
 
     public event Action<string, int> OnTextChanged;
-    public event Action<int, int> OnProgressChanged;
+    public event Action<double, int> OnProgressChanged;
   }
 }
